+++ src/main/java/io/metaloom/quadtree/Size.java
@@ -3,17 +3,33 @@ package io.metaloom.quadtree;
 import io.metaloom.quadtree.impl.SizeImpl;
 import io.metaloom.quadtree.point.Point;
 
+/**
+ * Dimensional representation for quad trees and their elements.
+ */
 public interface Size {
 
-	int height();
+	long height();
 
-	int width();
+	long width();
 
+	/**
+	 * Create a new size which uses the coordinates of the point as dimensional references.
+	 * 
+	 * @param point
+	 * @return
+	 */
 	static Size of(Point point) {
 		return new SizeImpl(point.x(), point.y());
 	}
 
-	static Size of(int width, int height) {
+	/**
+	 * Create a new size.
+	 * 
+	 * @param width
+	 * @param height
+	 * @return
+	 */
+	static Size of(long width, long height) {
 		return new SizeImpl(width, height);
 	}
 }
